Fix XAudio version to pass the new tests
