+++ MonoGame.Framework/Android/AndroidGamePlatform.cs
@@ -72,8 +72,6 @@ namespace Microsoft.Xna.Framework
                 _initialized = true;
             }
 
-            DynamicSoundEffectInstanceManager.UpdatePlayingInstances();
-
             return true;
         }
