+++ MonoGame.Framework/Game.cs
@@ -664,6 +664,7 @@ namespace Microsoft.Xna.Framework
                 // playing sounds to see if they've stopped,
                 // and return them back to the pool if so.
                 SoundEffectInstancePool.Update();
+                DynamicSoundEffectInstanceManager.UpdatePlayingInstances();
 
                 Update(gameTime);
