+++ MonoGame.Framework/iOS/iOSGamePlatform.cs
@@ -246,7 +246,6 @@ namespace Microsoft.Xna.Framework
         {
     		// Update our OpenAL sound buffer pools
     		soundControllerInstance.Update();
-    		DynamicSoundEffectInstanceManager.UpdatePlayingInstances();
 
             if (IsPlayingVideo)
                 return false;
